Refactor data type for latitude/longitude

+++ TASI.Backend/Domain/Suppliers/Dtos/EditSupplierDto.cs
@@ -6,7 +6,7 @@ namespace TASI.Backend.Domain.Suppliers.Dtos
     {
         public string? Name { get; set; }
         public string? Address { get; set; }
-        public long? Latitude { get; set; }
-        public long? Longitude { get; set; }
+        public double? Latitude { get; set; }
+        public double? Longitude { get; set; }
     }
 }
